[ etc ] move theme provider files to theme folder
