+++ agoda_toy/src/styles/theme/styled.d.ts
